pyproject.toml
[project]
name = "suri"
version = "0.0.1"
description = "Statistical Understanding & Reasoning Interface — Data Agent PoC"
requires-python = ">=3.12"
dependencies = [
  "streamlit>=1.40",
  "python-dotenv>=1.0",
  "faker>=40.15.0",
  "psycopg[binary]>=3.3.3",
  "mcp[cli]>=1.27.0",
  "sqlglot>=30.6.0",
]

[tool.uv]
# D-6 이후 추가 예정:
# langchain, langgraph, langchain-anthropic,
# mcp, psycopg[binary], statsmodels, pandas
